requirements.txt
google-api-python-client>=2.184.0
google-auth>=2.41.1
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.2
requests>=2.31.0
beautifulsoup4>=4.12.0
